pyproject.toml
[project]
name = "nutrition-fact-reader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiohttp-retry==2.9.1",
    "aiosignal==1.3.2",
    "aiostream==0.5.2",
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anthropic==0.46.0",
    "anyio==4.9.0",
    "arpeggio==2.0.2",
    "asyncpg==0.30.0",
    "attrs==25.3.0",
    "barcode>=1.0.4",
    "beautifulsoup4==4.13.4",
    "blinker==1.9.0",
    "cachetools==5.5.2",
    "caliper-reader==0.4.1",
    "cel-python==0.2.0",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "cfgv==3.4.0",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "contourpy==1.3.2",
    "cryptography==45.0.4",
    "cycler==0.12.1",
    "deprecated==1.2.18",
    "deprecation==2.1.0",
    "dill==0.4.0",
    "diskcache==5.6.3",
    "distlib==0.3.9",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "einops==0.8.1",
    "email-validator==2.2.0",
    "fastapi==0.115.12",
    "fastapi-cli==0.0.7",
    "filelock==3.18.0",
    "filetype==1.2.0",
    "fonttools==4.58.0",
    "frozenlist==1.6.0",
    "fsspec==2025.5.1",
    "ftfy==6.3.1",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "google-auth==2.40.3",
    "google-genai==1.21.1",
    "greenlet==3.2.2",
    "grpcio==1.71.0",
    "grpcio-tools==1.71.0",
    "h11==0.16.0",
    "hatchet==1.4.0",
    "hatchet-sdk==1.10.1",
    "hf-xet==1.1.5",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "huggingface-hub==0.33.0",
    "identify==2.6.12",
    "idna==3.10",
    "img2pdf==0.6.1",
    "iniconfig==2.1.0",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "jmespath==1.0.1",
    "joblib==1.5.1",
    "jsonschema==4.24.0",
    "jsonschema-specifications==2025.4.1",
    "kiwisolver==1.4.8",
    "lark==0.12.0",
    "llama-cpp-python==0.3.9",
    "llava==0.0.1.dev0",
    "lxml==5.4.0",
    "markdown-it-py==3.0.0",
    "markdown2==2.5.3",
    "markdownify==0.13.1",
    "marker-pdf==1.7.5",
    "markupsafe==3.0.2",
    "matplotlib==3.10.3",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "multidict==6.4.3",
    "multiprocess==0.70.18",
    "narwhals==1.43.1",
    "networkx==3.5",
    "nodeenv==1.9.1",
    "numpy==1.26.4",
    "nvidia-cublas-cu12==12.6.4.1",
    "nvidia-cuda-cupti-cu12==12.6.80",
    "nvidia-cuda-nvrtc-cu12==12.6.77",
    "nvidia-cuda-runtime-cu12==12.6.77",
    "nvidia-cudnn-cu12==9.5.1.17",
    "nvidia-cufft-cu12==11.3.0.4",
    "nvidia-cufile-cu12==1.11.1.6",
    "nvidia-curand-cu12==10.3.7.77",
    "nvidia-cusolver-cu12==11.7.1.2",
    "nvidia-cusparse-cu12==12.5.4.2",
    "nvidia-cusparselt-cu12==0.6.3",
    "nvidia-nccl-cu12==2.26.2",
    "nvidia-nvjitlink-cu12==12.6.85",
    "nvidia-nvtx-cu12==12.6.77",
    "ocrmypdf==16.10.2",
    "openai==1.79.0",
    "opencv-python-headless==4.11.0.86",
    "packaging==25.0",
    "pandas==2.2.3",
    "pdfminer-six==20250506",
    "pdftext==0.6.3",
    "pi-heif==0.22.0",
    "pikepdf==9.9.0",
    "pillow==10.4.0",
    "platformdirs==4.3.8",
    "pluggy==1.6.0",
    "pre-commit==4.2.0",
    "prometheus-client==0.21.1",
    "propcache==0.3.1",
    "protobuf==5.29.4",
    "psycopg2-binary==2.9.10",
    "pyarrow==20.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pycubexr==2.0.1",
    "pydantic==2.11.4",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.9.1",
    "pydeck==0.9.1",
    "pydot==4.0.0",
    "pygments==2.19.1",
    "pyparsing==3.2.3",
    "pypdfium2==4.30.0",
    "pytesseract==0.3.13",
    "pytest==8.3.5",
    "pytest-asyncio==0.26.0",
    "python-barcode[images]>=0.15.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "pyzbar>=0.1.9",
    "qrcode[pil]>=8.2",
    "rapidfuzz==3.13.0",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.4",
    "rich==14.0.0",
    "rich-toolkit==0.14.6",
    "rpds-py==0.25.1",
    "rsa==4.9.1",
    "safetensors==0.5.3",
    "scikit-learn==1.7.0",
    "scipy==1.16.0",
    "setuptools==80.7.1",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sqlalchemy==2.0.41",
    "sqlmodel==0.0.24",
    "starlette==0.46.2",
    "streamlit==1.46.0",
    "surya-ocr==0.14.6",
    "sympy==1.14.0",
    "tenacity==8.5.0",
    "textx==4.2.2",
    "threadpoolctl==3.6.0",
    "tokenizers==0.21.1",
    "toml==0.10.2",
    "torch==2.7.1",
    "tornado==6.5.1",
    "tqdm==4.67.1",
    "transformers==4.52.4",
    "triton==3.3.1",
    "typer==0.16.0",
    "types-python-dateutil==2.9.0.20250516",
    "types-pyyaml==6.0.12.20250516",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "uvloop==0.21.0",
    "virtualenv==20.31.2",
    "watchdog==6.0.0",
    "watchfiles==1.0.5",
    "wcwidth==0.2.13",
    "websockets==15.0.1",
    "wrapt==1.17.2",
    "yarl==1.20.0",
]
